Formatted employee-web module

+++ liferay/modules/employee-web/src/main/java/com/employee/portlet/DeleteEmployeeAction.java
@@ -17,17 +17,15 @@ import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 
-@Component(
-	property = { 
-		"javax.portlet.name =" + EmployeeWebPortletKeys.EMPLOYEEWEB,
-		"mvc.command.name=/deleteEmployee", 
-	}, service = MVCActionCommand.class)
+@Component(property = { "javax.portlet.name =" + EmployeeWebPortletKeys.EMPLOYEEWEB,
+		"mvc.command.name=/deleteEmployee", }, service = MVCActionCommand.class)
 
 public class DeleteEmployeeAction extends BaseMVCActionCommand {
 
-	private static Log log = LogFactoryUtil.getLog(DeleteEmployeeAction.class);
+	private static final Log log = LogFactoryUtil.getLog(DeleteEmployeeAction.class);
 
 	@Reference
 	CounterLocalService counterLocalService;
@@ -39,7 +37,8 @@ public class DeleteEmployeeAction extends BaseMVCActionCommand {
 	UserLocalService userLocalService;
 
 	@Override
-	protected void doProcessAction(ActionRequest actionRequest, ActionResponse actionResponse) throws Exception {
+	protected void doProcessAction(ActionRequest actionRequest, ActionResponse actionResponse) 
+			throws Exception {
 
 		log.info("Delete Action method is started");
 
@@ -49,17 +48,26 @@ public class DeleteEmployeeAction extends BaseMVCActionCommand {
 		long employeeId = ParamUtil.getLong(actionRequest, EmployeeWebPortletKeys.EMPLOYEE_ID, GetterUtil.DEFAULT_LONG);
 
 		try {
-			EmployeeDetail employee = employeedetailLocalService.getEmployeeDetail(employeeId);
-			String userEmail = employee.getEmail();
-			employeedetailLocalService.deleteEmployeeDetail(employeeId);
-			log.info("Employee data is deleted");
-			
-			User user = userLocalService.getUserByEmailAddress(companyId, userEmail);
-			long userId = user.getUserId();
-			userLocalService.deleteUser(userId);
-			log.info("User data is deleted");
+			if (Validator.isNotNull(employeeId)) {
+				EmployeeDetail employee = employeedetailLocalService.fetchEmployeeDetail(employeeId);
+				String userEmail = employee.getEmail();
+				employeedetailLocalService.deleteEmployeeDetail(employeeId);
+				log.info("Employee data is deleted");
+
+				User user = userLocalService.getUserByEmailAddress(companyId, userEmail);
+				long userId = user.getUserId();
+				userLocalService.deleteUser(userId);
+				log.info("User data is deleted");
+				
+				String url = themeDisplay.getURLPortal() + themeDisplay.getURLCurrent();
+				actionResponse.sendRedirect(url);
+				
+			} else {
+				log.error("Employee Id fetched as null");
+			}
+
 		} catch (Exception e) {
-			log.error("error in deleteActionCommand");
+			log.error("Error in deleteActionCommand: ", e);
 		}
 	}
 }
